requirements.txt
fastapi
uvicorn[standard]
pandas
pyarrow
fsspec
hf_transfer
openai
python-dotenv